requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs==22.2.0
autopep8==2.0.2
av==10.0.0
ax-platform==0.2.5.1
Babel==2.11.0
backcall==0.2.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
botorch==0.6.4
bpemb==0.3.4
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.33
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffimg==0.2.3
dill==0.3.4
distlib==0.3.6
docopt==0.6.2
docstring-parser==0.15
easydict==1.10
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.89.1
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.9.0
filterpy==1.4.5
fire==0.5.0
Flask==2.2.2
Flask-Babel==2.0.0
flatbuffers==2.0.7
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.6.4
gensim==4.2.0
gevent==22.10.2
geventhttpclient==2.0.2
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpytorch==1.8.1
gradio==3.17.1
greenlet==2.0.2
grpcio==1.43.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
humanfriendly==10.0
hydra-core==1.2.0
idna==3.4
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iopath==0.1.9
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonargparse==4.21.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-server==1.23.5
jupyter-ydoc==0.2.2
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.8.0
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
kaggle==1.5.13
kiwisolver==1.4.4
lanms-neo==1.0.2
lap==0.4.0
lightning-utilities==0.8.0
linkify-it-py==1.0.3
lmdb==1.3.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.12.2
munch==3.0.0
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnxconverter-common==1.13.0
onnxruntime==1.14.0
onnxruntime-gpu==1.14.0
onnxruntime-tools==1.7.0
onnxsim==0.4.17
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openpyxl==3.1.0
opt-einsum==3.3.0
optimum==1.7.1
orjson==3.8.5
packaging==23.0
paddle-bfloat==0.1.7
paddle2onnx==1.0.5
paddlefsl==1.1.0
paddlenlp==2.5.1
paddlepaddle==2.4.1
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.13
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.13.0
Polygon3==3.0.9.1
portalocker==2.7.0
premailer==3.10.0
pretrainedmodels==0.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.14.0
PyMuPDF==1.20.2
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.5
python-rapidjson==1.9
python-slugify==8.0.1
pytorch-lightning==1.9.5
pytorchvideo==0.1.5
PyTurboJPEG==1.7.0
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
rapidfuzz==2.13.7
rarfile==4.0
ray==1.13.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@e7cbea0bbd546167b9d1757b2cfeca3fbe26a33e
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.19.1
seqeval==1.2.2
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
sympy==1.10.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tf2onnx==1.13.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
trdg==1.8.0
tritonclient==2.30.0
typed-ast==1.5.4
typeguard==2.13.3
typer==0.7.0
typeshed-client==2.3.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
-e git+https://github.com/ultralytics/ultralytics.git@5065ca36a8499b43dd4b938ba73ddb98bc6e278f#egg=ultralytics
Unidecode==1.3.6
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
visualdl==2.4.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
x2paddle==1.4.0
xmltodict==0.13.0
xxhash==3.2.0
y-py==0.5.9
yacs==0.1.8
yarg==0.1.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.12.1
zope.event==4.6
zope.interface==5.5.2
